beauty1/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
Django==3.2.7
django-js-asset==1.2.2
django-mptt==0.13.4
idna==3.2
Pillow==8.4.0
pyTelegramBotAPI==4.0.1
pytz==2021.1
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.6
